+++ CHANGES.rst
@@ -26,6 +26,7 @@ v2.0.0 (xxxx-xx-xx)
 - Rename *private* utility function ``_get_item()`` to ``getitem()``. **breaking change**
 - Rename *private* utility function ``_set_item()`` to ``setitem()``. **breaking change**
 - Rename *private* utility function ``_deprecated()`` to ``deprecated()``. **breaking change**
+- Modify ``is_number()`` to return ``False`` for boolean ``True`` and ``False``. (**breaking change**)
 - Undeprecate ``tail()`` and make alias of ``rest()``.